+++ src/MainChainParams.nim
@@ -20,9 +20,9 @@ type ChainParams = object
   #Initial Blockchain Difficulty.
   BLOCK_DIFFICULTY: uint64
   #Initial Send Difficulty.
-  SEND_DIFFICULTY: uint32
+  SEND_DIFFICULTY: uint16
   #Initial Data Difficulty.
-  DATA_DIFFICULTY: uint32
+  DATA_DIFFICULTY: uint16
 
   NETWORK_PROTOCOL: uint
   NETWORK_ID: uint
